+++ cedar-policy-core/src/ast/ops.rs
@@ -89,6 +89,16 @@ pub enum BinaryOp {
     ///
     /// Arguments must have Set type
     ContainsAny,
+
+    /// Get a tag of an entity.
+    ///
+    /// First argument must have Entity type, second argument must have String type.
+    GetTag,
+
+    /// Does the given `expr` have the given `tag`?
+    ///
+    /// First argument must have Entity type, second argument must have String type.
+    HasTag,
 }
 
 impl std::fmt::Display for UnaryOp {
@@ -113,6 +123,8 @@ impl std::fmt::Display for BinaryOp {
             BinaryOp::Contains => write!(f, "contains"),
             BinaryOp::ContainsAll => write!(f, "containsAll"),
             BinaryOp::ContainsAny => write!(f, "containsAny"),
+            BinaryOp::GetTag => write!(f, "getTag"),
+            BinaryOp::HasTag => write!(f, "hasTag"),
         }
     }
 }