+++ cedar-policy-core/src/parser.rs
@@ -573,6 +573,7 @@ mod tests {
         );
     }
 
+    /// Tests parser+evaluator with relations `<`, `<=`, `>`, `&&`, `||`, `!=`
     #[test]
     fn interpret_relation() {
         let request = eval::test::basic_request();
@@ -608,6 +609,34 @@ mod tests {
             val.source_loc().unwrap().snippet(),
             Some("7 <= 7 && 4 != 5")
         );
+    }
+
+    /// Tests parser+evaluator with builtin methods `containsAll()`, `hasTag`, `getTag`
+    #[test]
+    fn interpret_methods() {
+        let request = eval::test::basic_request();
+        let entities = eval::test::basic_entities();
+        let exts = Extensions::none();
+        let evaluator = eval::Evaluator::new(request, &entities, exts);
+        // The below tests check not only that we get the expected `Value`, but
+        // that it has the expected source location.
+        // See note on this in the above test.
+
+        let src = r#"
+            [2, 3, "foo"].containsAll([3, "foo"])
+            && principal.hasTag(resource.getTag(context.cur_time))
+        "#;
+        let expr = parse_expr(src).unwrap();
+        assert_matches!(evaluator.interpret_inline_policy(&expr), Err(e) => {
+            expect_err(
+                src,
+                &miette::Report::new(e),
+                &ExpectedErrorMessageBuilder::error(r#"`test_entity_type::"test_resource"` does not have the tag `03:22:11`"#)
+                    .help(r#"`test_entity_type::"test_resource"` does not have any tags"#)
+                    .exactly_one_underline("resource.getTag(context.cur_time)")
+                    .build(),
+            );
+        });
     }
 
     #[test]