+++ cedar-policy-core/src/parser/cst_to_ast.rs
@@ -459,6 +459,10 @@ impl ast::UnreservedId {
                 .map(|arg| construct_method_contains_all(e, arg, loc.clone())),
             "containsAny" => extract_single_argument(args.into_iter(), "containsAny", loc)
                 .map(|arg| construct_method_contains_any(e, arg, loc.clone())),
+            "getTag" => extract_single_argument(args.into_iter(), "getTag", loc)
+                .map(|arg| construct_method_getTag(e, arg, loc.clone())),
+            "hasTag" => extract_single_argument(args.into_iter(), "hasTag", loc)
+                .map(|arg| construct_method_hasTag(e, arg, loc.clone())),
             _ => {
                 if EXTENSION_STYLES.methods.contains(self) {
                     let args = NonEmpty {
@@ -1514,7 +1518,10 @@ impl ast::Name {
         if self.0.path.is_empty() {
             let id = self.basename();
             if EXTENSION_STYLES.methods.contains(&id)
-                || matches!(id.as_ref(), "contains" | "containsAll" | "containsAny")
+                || matches!(
+                    id.as_ref(),
+                    "contains" | "containsAll" | "containsAny" | "getTag" | "hasTag"
+                )
             {
                 return Err(ToASTError::new(
                     ToASTErrorKind::FunctionCallOnMethod(self.basename()),
@@ -1796,6 +1803,12 @@ fn construct_method_contains_any(e0: ast::Expr, e1: ast::Expr, loc: Loc) -> ast:
     ast::ExprBuilder::new()
         .with_source_loc(loc)
         .contains_any(e0, e1)
+}
+fn construct_method_getTag(e0: ast::Expr, e1: ast::Expr, loc: Loc) -> ast::Expr {
+    ast::ExprBuilder::new().with_source_loc(loc).get_tag(e0, e1)
+}
+fn construct_method_hasTag(e0: ast::Expr, e1: ast::Expr, loc: Loc) -> ast::Expr {
+    ast::ExprBuilder::new().with_source_loc(loc).has_tag(e0, e1)
 }
 
 fn construct_ext_meth(n: UnreservedId, args: NonEmpty<ast::Expr>, loc: Loc) -> ast::Expr {