+++ cedar-policy/CHANGELOG.md
@@ -13,6 +13,8 @@ Starting with version 3.2.4, changes marked with a star (*) are _language breaki
 Cedar Language Version: TBD
 
 ### Added
+- Implemented [RFC 82](https://github.com/cedar-policy/rfcs/pull/82), adding
+  entity tags to the Cedar language (#1204, more coming)
 - Implemented [RFC 74](https://github.com/cedar-policy/rfcs/pull/74): A new experimental API (`compute_entity_manifest`)
   that provides the Entity Manifest: a data
   structure that describes what data is required to satisfy a